+++ app/(components)/AuthModal.tsx
@@ -1,19 +1,12 @@
 "use client"
 import React, { Fragment, useState } from "react"
-import Link from "next/link"
+import { Dialog, Menu, Tab, Transition } from "@headlessui/react"
+import { classNames } from "@/lib/classNames"
 import LoginForm from "@/app/(components)/LoginForm"
 import RegisterForm from "@/app/(components)/RegisterForm"
-import { useAuth } from "@/helpers/pocketbase"
-import useCurrentUser from "@/app/(components)/useCurrentUser"
-import { Menu, Transition, Dialog, Tab } from "@headlessui/react"
-import Image from "next/image"
-import { classNames } from "@/lib/classNames"
 
-const NavBar = () => {
-  const currentUser = useCurrentUser()
-  const { logout } = useAuth()
+const AuthModal = () => {
   let [isOpen, setIsOpen] = useState(false)
-
   function closeModal() {
     setIsOpen(false)
   }
@@ -25,65 +18,12 @@ const NavBar = () => {
   const authProps = {
     closeModal,
   }
+
   return (
     <>
-      <nav className={"flex justify-around items-center min-w-screen"}>
-        <ul className="flex items-center gap-4">
-          <li>
-            <Link href="/" className={"font-bold text-lg"}>
-              AI Project Creator POC
-            </Link>
-          </li>
-          <li>
-            <Link href="/about">About</Link>
-          </li>
-        </ul>
-        {currentUser ? (
-          <Menu as="div" className="relative">
-            <Menu.Button className="">
-              <Image
-                src={`https://pocketcreator.jimmymcbride.dev/api/files/_pb_users_auth_/ooelbyidysr4c9n/${currentUser?.avatar}`}
-                className="rounded-full"
-                width={32}
-                height={32}
-                alt="Avatar"
-              />
-            </Menu.Button>
-
-            <Transition
-              as={Fragment}
-              enter="transition ease-out duration-100"
-              enterFrom="transform opacity-0 scale-95"
-              enterTo="transform opacity-100 scale-100"
-              leave="transition ease-in duration-75"
-              leaveFrom="transform opacity-100 scale-100"
-              leaveTo="transform opacity-0 scale-95"
-            >
-              <Menu.Items className="absolute right-0 z-10 mt-2 w-56 origin-top-right rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
-                <div className="py-1">
-                  <Menu.Item>
-                    {({ active }) => (
-                      <button
-                        type="button"
-                        className={`${
-                          active ? "bg-gray-100 text-gray-900" : "text-gray-700"
-                        } block w-full px-4 py-2 text-left text-sm`}
-                        onClick={logout}
-                      >
-                        Sign out
-                      </button>
-                    )}
-                  </Menu.Item>
-                </div>
-              </Menu.Items>
-            </Transition>
-          </Menu>
-        ) : (
-          <div>
-            <button onClick={openModal}>Login Or Sign Up!</button>
-          </div>
-        )}
-      </nav>
+      <div>
+        <button onClick={openModal}>Login Or Sign Up!</button>
+      </div>
       <Transition appear show={isOpen} as={Fragment}>
         <Dialog as="div" className="relative z-10" onClose={closeModal}>
           <Transition.Child
@@ -165,4 +105,8 @@ const NavBar = () => {
   )
 }
 
-export default NavBar
+AuthModal.propTypes = {
+  // currentUser: PropTypes.object,
+}
+
+export default AuthModal
