+++ app/(components)/NameBoardButton.tsx
@@ -1,12 +1,38 @@
-import React, { Suspense } from "react"
+"use client"
+import React, { Suspense, useEffect, useState } from "react"
 
-const NameBoardButton = ({
-  data,
-  createTrelloBoard,
-}: {
-  data: string
-  createTrelloBoard: (data: string) => void
-}) => {
+const createTrelloBoard = (boardName: String) => {
+  console.log(`Board Name: ${boardName}`)
+  fetch(`/api/trello?boardName=${boardName}`, {
+    method: "POST",
+    body: JSON.stringify({ boardName }),
+  })
+    .then((r) => r.json())
+    .then((data) => console.log("Data: " + data.json()))
+    .catch((error) => console.error(error))
+}
+
+const NameBoardButton = () => {
+  const [data, setData] = useState("")
+  const [isLoading, setLoading] = useState(false)
+
+  useEffect(() => {
+    setLoading(true)
+    fetch("/api/openai", {
+      method: "POST",
+      body: JSON.stringify({ project: "Spotify clone" }),
+    })
+      .then((res) => res.json())
+      .then((data) => {
+        console.log(`Data Name: ${data.name}`)
+        setData(data.name)
+        setLoading(false)
+      })
+      .catch((error) => {
+        console.error(error)
+        setLoading(false)
+      })
+  }, [])
   return (
     <div className={"flex flex-col justify-center items-center"}>
       <button className={"mt-4 btn"} onClick={() => createTrelloBoard(data)}>
