+++ app/dashboard/page.tsx
@@ -1,0 +1,7 @@
+export default function Dashboard() {
+  return (
+    <>
+      <div>Hello, world!</div>
+    </>
+  )
+}
