+++ app/loading.tsx
@@ -1,0 +1,5 @@
+import Spinner from "@/components/Spinner"
+
+export default function Loading() {
+  return <Spinner />
+}
