+++ lib/pb.ts
@@ -1,0 +1,3 @@
+import PocketBase from "pocketbase"
+
+export const pb = new PocketBase("https://pocketcreator.jimmymcbride.dev")
